# backend/docs/swagger.yaml
basePath: /
definitions:
...
  installer.CompleteInstallationRequest:
    properties:
      device_id:
        type: string
      inverter_brand:
        type: string
      plant_capacity_kw:
        type: number
    required:
    - device_id
    type: object
  inventory.InventoryItem:
    properties:
      brand:
        type: string
      category:
        description: e.g., "Solar Panel", "Inverter", "Battery", "Mounting"
        type: string
      created_at:
        type: string
      description:
        type: string
      gallery:
        description: JSON string of additional image URLs
        type: string
      id:
        type: string
      image_url:
        description: Primary image
        type: string
      is_active:
        type: boolean
      model:
        type: string
      name:
        type: string
      price:
        type: number
      specs:
        description: JSON string of technical specifications
        type: string
      stock_qty:
        type: integer
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API documentation for the Solar Energy Management System.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Solar Energy Management System API
  version: "1.0"
paths:
...
  /installer/jobs:
    get:
      consumes:
      - application/json
      description: Get a list of jobs available for installation
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get available installation jobs
      tags:
      - Installer
  /installer/jobs/{id}/complete:
    post:
      consumes:
      - application/json
      description: Mark an installation job as complete and link a device
      parameters:
      - description: User ID (Job ID)
        in: path
        name: id
        required: true
        type: string
      - description: Completion details
        in: body
        name: completion
        required: true
        schema:
          $ref: '#/definitions/installer.CompleteInstallationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Complete installation
      tags:
      - Installer
...
  /inventory:
    get:
      consumes:
      - application/json
      description: Get a list of all inventory items
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/inventory.InventoryItem'
              type: array
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all inventory items
      tags:
      - Inventory
    post:
      consumes:
      - application/json
      description: Create a new item in the inventory (Admin only)
      parameters:
      - description: Inventory Item
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/inventory.InventoryItem'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create inventory item
      tags:
      - Inventory
  /inventory/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an inventory item (Admin only)
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete inventory item
      tags:
      - Inventory
    get:
      consumes:
      - application/json
      description: Get details of a specific inventory item
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/inventory.InventoryItem'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get inventory item
      tags:
      - Inventory
    put:
      consumes:
      - application/json
      description: Update an inventory item (Admin only)
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: string
      - description: Item update data
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/inventory.InventoryItem'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update inventory item
      tags:
      - Inventory
...
  /user/notifications:
    get:
      consumes:
      - application/json
      description: Get current user's notifications with pagination
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get user notifications
      tags:
      - Notifications
  /user/notifications/{id}/read:
    put:
      consumes:
      - application/json
      description: Mark a specific notification as read
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Mark as read
      tags:
      - Notifications
  /user/notifications/preferences:
    get:
      consumes:
      - application/json
      description: Get notification preferences for current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get preferences
      tags:
      - Notifications
  /user/notifications/read-all:
    put:
      consumes:
      - application/json
      description: Mark all notifications as read for current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Mark all as read
      tags:
      - Notifications
  /user/notifications/unread-count:
    get:
      consumes:
      - application/json
      description: Get count of unread notifications for current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: integer
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get unread count
      tags:
      - Notifications
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"